% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyclostratigraphics_gui.R
\name{cyclostratigraphics_gui}
\alias{cyclostratigraphics_gui}
\title{cyclostratigraphics_gui}
\usage{
cyclostratigraphics_gui()
}
\description{
GUI to help in the spectral and wavelet analysis of stratigraphic signals.
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
